+++ src/main/java/com/iexec/core/Application.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020-2023 IEXEC BLOCKCHAIN TECH
+ * Copyright 2020-2025 IEXEC BLOCKCHAIN TECH
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,18 +16,26 @@
 
 package com.iexec.core;
 
-import io.changock.runner.spring.v5.config.EnableChangock;
+import io.mongock.driver.mongodb.springdata.v4.SpringDataMongoV4Driver;
+import io.mongock.runner.springboot.EnableMongock;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.context.properties.ConfigurationPropertiesScan;
+import org.springframework.context.annotation.Bean;
+import org.springframework.data.mongodb.core.MongoTemplate;
 
-@EnableChangock
+@EnableMongock
 @SpringBootApplication
 @ConfigurationPropertiesScan
 public class Application {
 
     public static void main(String[] args) {
         SpringApplication.run(Application.class, args);
+    }
+
+    @Bean
+    public SpringDataMongoV4Driver mongockDriver(MongoTemplate mongoTemplate) {
+        return SpringDataMongoV4Driver.withDefaultLock(mongoTemplate);
     }
 
 }