+++ src/main/java/com/iexec/core/chain/adapter/BlockchainAdapterClientConfig.java
@@ -23,13 +23,11 @@ import feign.Logger;
 import lombok.Data;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.boot.context.properties.ConstructorBinding;
 import org.springframework.context.annotation.Bean;
 
 import java.time.Duration;
 
 @Data
-@ConstructorBinding
 @ConfigurationProperties(prefix = "blockchain-adapter")
 public class BlockchainAdapterClientConfig {