+++ src/main/java/com/iexec/core/configuration/ResultRepositoryConfiguration.java
@@ -23,10 +23,8 @@ import lombok.Value;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.boot.context.properties.ConstructorBinding;
 
 @Value
-@ConstructorBinding
 @ConfigurationProperties(prefix = "result-repository")
 @Slf4j
 public class ResultRepositoryConfiguration {